app/requirements.txt
psycopg2-binary
logzero